# docs/swagger.yaml
definitions:
  models.ErrorResponse:
    properties:
      remark:
        type: string
    type: object
  models.LoginResponse:
    properties:
      id:
        example: 1
        type: integer
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  models.RegisterRequest:
    properties:
      nama:
        type: string
      nik:
        type: string
      no_hp:
        type: string
    type: object
  models.RegisterResponse:
    properties:
      no_rekening:
        type: string
    type: object
  models.SaldoResponse:
    properties:
      saldo:
        type: integer
    type: object
  models.TabungTarikRequest:
    properties:
      no_rekening:
        type: string
      nominal:
        type: integer
    type: object
  models.User:
    properties:
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      password:
        example: securepassword123
        type: string
      username:
        example: johndoe
        type: string
    type: object
  models.UserResponse:
    properties:
      created_at:
        example: "2023-06-15T14:30:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      id:
        example: 1
        type: integer
      is_admin:
        example: false
        type: boolean
      updated_at:
        example: "2023-06-15T14:30:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/accounts/register:
    post:
      consumes:
      - application/json
      description: Register a new customer account with NIK, phone number, and name
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new account
      tags:
      - accounts
  /api/accounts/saldo/{no_rekening}:
    get:
      description: Retrieve current balance of an account
      parameters:
      - description: Account number
        in: path
        name: no_rekening
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaldoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get account balance
      tags:
      - accounts
  /api/accounts/tabung:
    post:
      consumes:
      - application/json
      description: Deposit money into an account
      parameters:
      - description: Deposit details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TabungTarikRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaldoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Deposit money
      tags:
      - transactions
  /api/accounts/tarik:
    post:
      consumes:
      - application/json
      description: Withdraw money from an account
      parameters:
      - description: Withdrawal details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TabungTarikRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SaldoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Withdraw money
      tags:
      - transactions
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Login user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the input payload
      parameters:
      - description: Register user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Register a new user
      tags:
      - auth
  /api/profile:
    get:
      consumes:
      - application/json
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user profile
      tags:
      - auth
swagger: "2.0"